[Enhancement] Split cast/agg template-cartesian TUs to cut compile time
Several expression/aggregate factory TUs were dominated by large 2-D template
cartesians (arg-type x ret-type / from-type x to-type), making them among the
slowest single translation units in the BE build. Split each, behavior-preservingly,
across parallel TUs: move the classes/templates/macros into a shared _tpl.hpp and
distribute the cartesian over group .cpp files, with the original file routing by
one type dimension.

- aggregate_resolver_avg: split out sum_map -> aggregate_resolver_sum_map{1,2,3}.cpp
  (641s -> 201 + 150/154/117s)
- case_expr: result-type cartesian -> case_expr_g{1,2,3}.cpp
  (536s -> 14 + 203/176/157s)
- aggregate_resolver_minmaxany: split out max_by/min_by ->
  aggregate_resolver_maxminby{1,2,3}.cpp (346s -> 187 + 90/90/76s)
- cast_expr: primitive from-type x to-type cartesian -> cast_expr_g{1,2,3}.cpp
  (258s -> 79 + 87/88/72s)

Behavior-preserving (identical type coverage). The wall-clock pole of each TU drops
~60%; total CPU rises slightly from per-group header re-parse. Isolated compile times
measured without PCH.
