+++ be/src/exprs/CMakeLists.txt
@@ -26,9 +26,15 @@ set(EXPR_FILES
   agg/factory/aggregate_factory.cpp
   agg/factory/aggregate_resolver_approx.cpp
   agg/factory/aggregate_resolver_avg.cpp
+  agg/factory/aggregate_resolver_sum_map1.cpp
+  agg/factory/aggregate_resolver_sum_map2.cpp
+  agg/factory/aggregate_resolver_sum_map3.cpp
   agg/factory/aggregate_resolver_boolean.cpp
   agg/factory/aggregate_resolver_hypothesis_testing.cpp
   agg/factory/aggregate_resolver_minmaxany.cpp
+  agg/factory/aggregate_resolver_maxminby1.cpp
+  agg/factory/aggregate_resolver_maxminby2.cpp
+  agg/factory/aggregate_resolver_maxminby3.cpp
   agg/factory/aggregate_resolver_others.cpp
   agg/factory/aggregate_resolver_sumcount.cpp
   agg/factory/aggregate_resolver_distinct.cpp
@@ -53,9 +59,15 @@ set(EXPR_FILES
   bitmap_functions.cpp
   cast_nested.cpp
   cast_expr.cpp
+  cast_expr_g1.cpp
+  cast_expr_g2.cpp
+  cast_expr_g3.cpp
   cast_expr_array.cpp
   cast_expr_struct.cpp
   case_expr.cpp
+  case_expr_g1.cpp
+  case_expr_g2.cpp
+  case_expr_g3.cpp
   cast_expr_json.cpp
   cast_expr_map.cpp
   chunk_predicate_evaluator.cpp
