+++ be/src/exprs/agg/factory/aggregate_factory.cpp
@@ -51,7 +51,13 @@ AggregateFunctionPtr resolve_non_builtin_aggregate_function(TFunctionBinaryType:
 
 AggregateFuncResolver::AggregateFuncResolver() {
     register_avg();
+    register_sum_map1();
+    register_sum_map2();
+    register_sum_map3();
     register_minmaxany();
+    register_maxminby1();
+    register_maxminby2();
+    register_maxminby3();
     register_bitmap();
     register_sumcount();
     register_distinct();
