+++ be/src/exprs/agg/factory/aggregate_resolver_minmaxany.cpp
@@ -101,48 +101,12 @@ struct MaxNDispatcher {
     }
 };
 
-template <LogicalType ret_type, bool is_max_by>
-struct MaxMinByDispatcherInner {
-    template <LogicalType arg_type>
-    void operator()(AggregateFuncResolver* resolver) {
-        if constexpr ((lt_is_aggregate<arg_type> || lt_is_json<arg_type>)&&(
-                              lt_is_aggregate<ret_type> || lt_is_json<ret_type> || lt_is_collection<ret_type>)) {
-            if constexpr (is_max_by) {
-                resolver->add_aggregate_mapping_notnull<arg_type, ret_type>(
-                        "max_by", true, AggregateFactory::MakeMaxByAggregateFunction<arg_type, false>());
-                resolver->add_aggregate_mapping_notnull<arg_type, ret_type>(
-                        "max_by_v2", true, AggregateFactory::MakeMaxByAggregateFunction<arg_type, true>());
-            } else {
-                resolver->add_aggregate_mapping_notnull<arg_type, ret_type>(
-                        "min_by", true, AggregateFactory::MakeMinByAggregateFunction<arg_type, false>());
-                resolver->add_aggregate_mapping_notnull<arg_type, ret_type>(
-                        "min_by_v2", true, AggregateFactory::MakeMinByAggregateFunction<arg_type, true>());
-            }
-        }
-    }
-};
-
-template <bool is_max_by>
-struct MaxMinByDispatcher {
-    template <LogicalType lt>
-    void operator()(AggregateFuncResolver* resolver, LogicalType ret_type) {
-        type_dispatch_all(ret_type, MaxMinByDispatcherInner<lt, is_max_by>(), resolver);
-    }
-};
-
 void AggregateFuncResolver::register_minmaxany() {
     auto minmax_types = aggregate_types();
     minmax_types.push_back(TYPE_JSON);
     minmax_types.push_back(TYPE_ARRAY);
     minmax_types.push_back(TYPE_STRUCT);
     minmax_types.push_back(TYPE_MAP);
-    for (auto ret_type : minmax_types) {
-        for (auto arg_type : minmax_types) {
-            type_dispatch_all(arg_type, MaxMinByDispatcher<true>(), this, ret_type);
-            type_dispatch_all(arg_type, MaxMinByDispatcher<false>(), this, ret_type);
-        }
-    }
-
     for (auto type : minmax_types) {
         type_dispatch_all(type, MinMaxAnyDispatcher(), this);
     }
