+++ be/src/exprs/agg/factory/aggregate_resolver_sum_map3.cpp
@@ -1,0 +1,26 @@
+// Copyright 2021-present StarRocks, Inc. All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include "exprs/agg/factory/aggregate_resolver_sum_map.hpp"
+
+namespace starrocks {
+
+// String / date / and remaining aggregate key types (non-aggregate ones are no-ops via the
+// lt_is_aggregate guard in SumMapDispatcher).
+void AggregateFuncResolver::register_sum_map3() {
+    register_sum_map_for_keys<TYPE_VARCHAR, TYPE_CHAR, TYPE_DATE, TYPE_DATETIME, TYPE_HLL, TYPE_OBJECT, TYPE_ARRAY>(
+            this);
+}
+
+} // namespace starrocks
