+++ be/src/exprs/case_expr_g3.cpp
@@ -1,0 +1,37 @@
+// Copyright 2021-present StarRocks, Inc. All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include "exprs/case_expr_tpl.hpp"
+
+namespace starrocks {
+
+// One third of the VectorizedCaseExpr<When, Result> result-type cartesian, split out so the
+// ~550 template instantiations compile across parallel TUs instead of one 536s file.
+Expr* case_from_thrift_group3(const starrocks::TExprNode& node, LogicalType resultType, LogicalType whenType) {
+    switch (resultType) {
+        CASE_RESULT_TYPE(TYPE_JSON)
+        CASE_RESULT_TYPE(TYPE_VARBINARY)
+        CASE_RESULT_TYPE(TYPE_VARIANT)
+        CASE_RESULT_TYPE(TYPE_STRUCT)
+        CASE_RESULT_TYPE(TYPE_MAP)
+        CASE_RESULT_TYPE(TYPE_ARRAY)
+        CASE_RESULT_TYPE(TYPE_OBJECT)
+        CASE_RESULT_TYPE(TYPE_HLL)
+        CASE_RESULT_TYPE(TYPE_PERCENTILE)
+    default:
+        return nullptr;
+    }
+}
+
+} // namespace starrocks
